==> 02_DGE.Rmd <==
---
title: "DGE"
author: "tobiasst"
date: '`r format(Sys.time(), "%d %B, %Y")`'
output:
  html_document:
    df_print: paged
    code_folding: hide
---

```{r init, message=F, warning=F}
library(DESeq2)
library(parallel)
library(pheatmap)
```

### HD vesus control in NSC 

```{r}
se <- readRDS("se.rds")
dds <- DESeqDataSet(se, design = ~ 0+group)
dds <- DESeq(dds)
resultsNames(dds)
res <- results(dds, contrast=c("group","HD_NSC","C116_NSC"))
summary(res)
saveRDS(res, "res_HD_NSC_v_C116_NSC.rds")

## fits suppl. table 
rf <- data.frame(res[order(res$padj),])
rf$symbol <- rowData(dds)$symbol

## differentially expressed genes
degf <- rf[!is.na(rf$padj) & rf$padj<0.1,]
res <- res[order(res$stat, decreasing = T),]

```


## top 50 upregulated genes

```{r heatmap, fig.height=8, fig.width=6}

flt <- !is.na(res$padj) & res$padj < 0.1 & res$stat > 0
upregulated <- rownames(res)[flt][1:50]

mat <- t(scale(t(log2(assays(se)$tpms[upregulated,]+0.01))))
rownames(mat) <- rowData(se[upregulated,])$symbol

ann <- data.frame(cell_type=se$cell_type, region=se$region,row.names = colnames(mat))
p <- pheatmap(mat, annotation_col = ann, scale = "row", cluster_rows=T)
```


```{r}
sessionInfo()
```
