==> 04_GSEA.Rmd <==
---
title: "GSEA"
author: "tobiasst"
date: '`r format(Sys.time(), "%d %B, %Y")`'
output:
  html_document:
    df_print: paged
    toc: true 
    toc_depth: 3
    number_sections: true  ## if you want number sections at each table header
    code_folding: hide
    css: share/custom.css

---

gene sets downloaded from 
http://bioinf.wehi.edu.au/software/MSigDB/index.html

```{r init, message=F, warning=F}
library(SummarizedExperiment)
library(fgsea)
library(org.Hs.eg.db)
library(pheatmap)
```

ranking on the statistic (alternative log2FC)

```{r getRanks, message=F, warning=F, error=F}
se <- readRDS("se.rds")
results <- readRDS("res_HD_NSC_v_C116_NSC.rds")

ranks <- results$stat
names(ranks) <- rowData(se)$entrez
ranks <- na.omit(ranks)
ranks <- ranks[!duplicated(names(ranks))]

```

## Hallmark gene sets 

```{r, message=F, fig.width=10, fig.height=10}
load("share/human_H_v5p2.rdata")
pathways <- Hs.H
fgseaRes <- fgsea(pathways, ranks, minSize=15, maxSize = 500, nperm=1000)
topPathwaysUp <- fgseaRes[ES > 0][head(order(pval), n=10), pathway]
topPathwaysDown <- fgseaRes[ES < 0][head(order(pval), n=10), pathway]
topPathways <- c(topPathwaysUp, rev(topPathwaysDown))
plotGseaTable(pathways[topPathways], ranks, fgseaRes, 
              gseaParam = 0.5)

```


### unfolded protein response

```{r, fig.height=15, fig.width=7}
eids <- pathways[["HALLMARK_UNFOLDED_PROTEIN_RESPONSE"]]

mat <- t(scale(t(log2(assays(se)$tpms[rowData(se)$entrez %in% eids,]+0.01))))
rownames(mat) <- rowData(se[rowData(se)$entrez %in% eids])$symbol

ann <- data.frame(cell_type=se$cell_type, region=se$region,row.names = colnames(mat))
p <- pheatmap(mat, annotation_col = ann, scale = "row", cluster_rows=T)

```

## GO gene sets

```{r, message=F, fig.width=10, fig.height=10}

load("share/human_c5_v5p2.rdata")
pathways <- Hs.c5
fgseaRes <- fgsea(pathways, ranks, minSize=15, maxSize = 500, nperm=1000)
topPathwaysUp <- fgseaRes[ES > 0][head(order(pval), n=10), pathway]
topPathwaysDown <- fgseaRes[ES < 0][head(order(pval), n=10), pathway]
topPathways <- c(topPathwaysUp, rev(topPathwaysDown))
plotGseaTable(pathways[topPathways], ranks, fgseaRes, 
              gseaParam = 0.5)

```

## curated gene sets

```{r, message=F, fig.width=10, fig.height=10}

load("share/human_c2_v5p2.rdata")
pathways <- Hs.c2
fgseaRes <- fgsea(pathways, ranks, minSize=15, maxSize = 500, nperm=1000)
topPathwaysUp <- fgseaRes[ES > 0][head(order(pval), n=10), pathway]
topPathwaysDown <- fgseaRes[ES < 0][head(order(pval), n=10), pathway]
topPathways <- c(topPathwaysUp, rev(topPathwaysDown))
plotGseaTable(pathways[topPathways], ranks, fgseaRes, 
              gseaParam = 0.5)

```


```{r}
sessionInfo()
```
